switch from btree to hashmap index
Switching from btree to hashmap since performance was pretty much the same and the hashmap index is much simpler.
